# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BookSummary {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cover_url?: string | null;
  chapter_count: number;
  created_at: string;
}

export interface ContentSummary {
  id: number;
  chapter_id: number;
  title: string;
  order_index: number;
  created_at: string;
}

export interface Chapter {
  id: number;
  book_id: number;
  title: string;
  order_index: number;
  contents: ContentSummary[];
  created_at: string;
}

export interface BookDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cover_url?: string | null;
  created_at: string;
  chapters: Chapter[];
}

export interface ContentDetail {
  id: number;
  chapter_id: number;
  title: string;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  video_url?: string | null;
  /** @nullable */
  video_id?: string | null;
  order_index: number;
  created_at: string;
  chapter_title: string;
  book_id: number;
  book_title: string;
}

export interface ButtonClickBody {
  user_id: string;
  button_label: string;
  content_id: number;
}

export interface VideoWatchBody {
  user_id: string;
  video_id: string;
  content_id: number;
  watched_seconds: number;
}

export interface ScrollDepthBody {
  user_id: string;
  content_id: number;
  max_scroll_percent: number;
}

export interface ButtonClickStat {
  button_label: string;
  content_id: number;
  content_title: string;
  count: number;
}

export interface VideoWatchStat {
  video_id: string;
  content_id: number;
  content_title: string;
  total_seconds: number;
  avg_seconds: number;
  sessions: number;
}

export interface ScrollDepthStat {
  content_id: number;
  content_title: string;
  avg_scroll_percent: number;
}

export interface AnalyticsSummary {
  buttonClicks: ButtonClickStat[];
  videoWatch: VideoWatchStat[];
  scrollDepth: ScrollDepthStat[];
}