# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ButtonClickStat } from "./buttonClickStat";
import type { ScrollDepthStat } from "./scrollDepthStat";
import type { VideoWatchStat } from "./videoWatchStat";

export interface AnalyticsSummary {
  buttonClicks: ButtonClickStat[];
  videoWatch: VideoWatchStat[];
  scrollDepth: ScrollDepthStat[];
}