# lib/api-zod/src/generated/types/bookDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Chapter } from "./chapter";

export interface BookDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cover_url?: string | null;
  created_at: string;
  chapters: Chapter[];
}