# lib/api-zod/src/generated/types/bookSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BookSummary {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cover_url?: string | null;
  chapter_count: number;
  created_at: string;
}