# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ContentSummary } from "./contentSummary";

export interface Chapter {
  id: number;
  book_id: number;
  title: string;
  order_index: number;
  contents: ContentSummary[];
  created_at: string;
}