# lib/api-zod/src/generated/types/contentDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentDetail {
  id: number;
  chapter_id: number;
  title: string;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  video_url?: string | null;
  /** @nullable */
  video_id?: string | null;
  order_index: number;
  created_at: string;
  chapter_title: string;
  book_id: number;
  book_title: string;
}