# lib/api-zod/src/generated/types/contentSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentSummary {
  id: number;
  chapter_id: number;
  title: string;
  order_index: number;
  created_at: string;
}