# lib/api-zod/src/generated/types/scrollDepthBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrollDepthBody {
  user_id: string;
  content_id: number;
  max_scroll_percent: number;
}