# lib/api-zod/src/generated/types/scrollDepthStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrollDepthStat {
  content_id: number;
  content_title: string;
  avg_scroll_percent: number;
}