# lib/api-zod/src/generated/types/videoWatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoWatchBody {
  user_id: string;
  video_id: string;
  content_id: number;
  watched_seconds: number;
}