# lib/api-zod/src/generated/types/videoWatchStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Content Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoWatchStat {
  video_id: string;
  content_id: number;
  content_title: string;
  total_seconds: number;
  avg_seconds: number;
  sessions: number;
}